llangchain-demo/requirements.txt
fastapi==0.112.1
langchain==0.2.14
langchain_community==0.2.12
langchain_openai==0.1.22
langchain_text_splitters==0.2.2
pydantic==2.8.2
python-dotenv==1.0.1
Requests==2.32.3
starlette==0.38.2
uvicorn==0.30.6
pymilvus==2.5.8
ollama==0.4.8
python-multipart==0.0.20
pdfplumber==0.11.6